Adding CPCA method and version upgrade to 1.4.0

+++ src/pca.h
@@ -43,7 +43,7 @@ double calcObjectDistance(matrix *m);
  * Calculate a principal component analysis
  */
 void PCA(matrix *mx,
-         size_t scaling,
+         int scaling,
          size_t npc,
          PCAMODEL *model,
          ssignal *s);
